Simple implementation of "winning" (not sent)
This commit attempts a simple implementation of the "winning" construct
in which it is assumed that the game sets "winning" based on a the
observed game state.  Since this can be done independently, the winning
indication is not transmitted.  In a game where the rules are not
completely determined (AnyCards in particular) "winning" (if supported
at all) must be set by a player.  In which case, it may need to be
transmitted.  Wait and see.  Issue #34